refactor(cu): split out loading process meta from loading process itself in readState
